# src/cleansia_customer_android/openapi/customer-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cleansia.Customer.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Payment/CreateOrder": {
      "post": {
        "tags": [
          "Payment"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.PaymentController.CreateOrder (Cleansia.Web.Customer)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Payment/webhook": {
      "post": {
        "tags": [
          "Payment"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.PaymentController.Webhook (Cleansia.Web.Customer)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateOrder_Command": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerEmail": {
            "type": "string",
            "nullable": true
          },
          "customerPhone": {
            "type": "string",
            "nullable": true
          },
          "customerAddress": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "savedAddressId": {
            "type": "string",
            "nullable": true
          },
          "selectedPackageIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "selectedServiceIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "rooms": {
            "type": "integer",
            "format": "int32"
          },
          "bathrooms": {
            "type": "integer",
            "format": "int32"
          },
          "extras": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "nullable": true
          },
          "cleaningDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentType": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "currencyId": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "language": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateOrder_Response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "confirmationCode": {
            "type": "string",
            "nullable": true
          },
          "stripeSessionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PaymentType": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Cash",
          "Card"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}